+++ runner.go
@@ -3,20 +3,19 @@ package goyek
 import (
 	"context"
 	"io"
-	"io/ioutil"
 	"time"
 )
 
-// Runner is used to run a Command.
-type Runner struct {
+// runner is used to run a Command.
+type runner struct {
 	Ctx         context.Context
 	TaskName    string
 	Output      io.Writer
 	ParamValues map[string]ParamValue
 }
 
-// RunResult contains the results of a Command run.
-type RunResult struct {
+// runResult contains the results of a Command run.
+type runResult struct {
 	failed   bool
 	skipped  bool
 	duration time.Duration
@@ -24,47 +23,29 @@ type RunResult struct {
 
 // Failed returns true if a command failed.
 // Failure can be caused by invocation of Error, Fail or related methods or a panic.
-func (r RunResult) Failed() bool {
+func (r runResult) Failed() bool {
 	return r.failed
 }
 
 // Skipped returns true if a command was skipped.
 // Skip is casused by invocation of Skip or related methods.
-func (r RunResult) Skipped() bool {
+func (r runResult) Skipped() bool {
 	return r.skipped
 }
 
-// Passed true if a command passed.
-// It means that it has not failed, nor skipped.
-func (r RunResult) Passed() bool {
-	return !r.failed && !r.skipped
-}
-
 // Duration returns the durations of the Command.
-func (r RunResult) Duration() time.Duration {
+func (r runResult) Duration() time.Duration {
 	return r.duration
 }
 
 // Run runs the command.
-func (r Runner) Run(command func(tf *TF)) RunResult {
-	ctx := context.Background()
-	if r.Ctx != nil {
-		ctx = r.Ctx
-	}
-	name := "no-name"
-	if r.TaskName != "" {
-		name = r.TaskName
-	}
-	writer := ioutil.Discard
-	if r.Output != nil {
-		writer = &syncWriter{Writer: r.Output}
-	}
-
-	finished := make(chan RunResult)
+func (r runner) Run(command func(tf *TF)) runResult {
+	finished := make(chan runResult)
 	go func() {
+		writer := &syncWriter{Writer: r.Output}
 		tf := &TF{
-			ctx:         ctx,
-			name:        name,
+			ctx:         r.Ctx,
+			name:        r.TaskName,
 			writer:      writer,
 			paramValues: r.ParamValues,
 		}
@@ -73,7 +54,7 @@ func (r Runner) Run(command func(tf *TF)) RunResult {
 			if r := recover(); r != nil {
 				tf.Errorf("panic: %v", r)
 			}
-			result := RunResult{
+			result := runResult{
 				failed:   tf.failed,
 				skipped:  tf.skipped,
 				duration: time.Since(from),